refactor(codex-rules): split rules engine loader
