+++ packages/omo-codex/plugin/components/rules/src/rules/constants.ts
@@ -28,10 +28,7 @@ export const PROJECT_RULE_SUBDIRS: ReadonlyArray<readonly [string, string]> = [
 /**
  * Single-file project rules (always apply, frontmatter optional).
  */
-export const PROJECT_SINGLE_FILES: readonly string[] = [
-	".github/copilot-instructions.md",
-	"CONTEXT.md",
-];
+export const PROJECT_SINGLE_FILES: readonly string[] = [".github/copilot-instructions.md", "CONTEXT.md"];
 
 /**
  * User-home rule directories.
